Reduce size of RenderText from 128 to 120 bytes on 64-bit https://bugs.webkit.org/show_bug.cgi?id=262655
Reviewed by Simon Fraser.

Reduce size of RenderText from 128 to 120 bytes on 64-bit by leveraging Markable.

* Source/WTF/wtf/Markable.h:
(WTF::FloatMarkableTraits::isEmptyValue):
(WTF::FloatMarkableTraits::emptyValue):
(WTF::Markable::value_or const):
* Source/WebCore/platform/graphics/FontSizeAdjust.h:
(WebCore::FloatMarkableTraits::isEmptyValue): Deleted.
(WebCore::FloatMarkableTraits::emptyValue): Deleted.
* Source/WebCore/rendering/RenderText.cpp:
(WebCore::RenderText::RenderText):
* Source/WebCore/rendering/RenderText.h:

Canonical link: https://commits.webkit.org/268887@main
